Fixed a bug where recording would never start
I'm so glad I copied all the console output from when we were playing a real wingman match. When recording is about to begin, this program uses the `status` command to find out what the map is named. The RegEX I made to grab the map name was
```
^map\ +: ([a-zA-Z_]+) at:.*$
```
because the line usually looks like `map     : de_vertigo at: -1443 x, -1517 y, 11524 z`. _However_, it turns out that you don't normally get the coordinates when not connected to a server so that map line looks like this when you're connected to an actual server: `map     : de_vertigo`. What's weird is I can't get this expression to match either of these anyways because of the beginning and end anchors unless I use the multiline flag but I don't even know if JS has the multiline flag enabled by default.

Also, I should've used `\s+` instead of `\ +` because that's kinda nasty but to be fair I was only looking for a space and not _any_ whitespace character. With all this in mind, the new (barely changed) RegEX becomes...
```
map\s+: ([a-zA-Z_]+).*
```
This is a certified bruh moment. I'm gonna make my promises timeout so that I'll know when something like this is happening again.

tl;dr: A RegEX failed to match a line that DemoRecordingHelper was supposed to read before it started recording a demo.