Use a proper logging solution + make promises for console output time out
Having promises that wait for console output is great but sometimes the program seems like it's just not working and this is because a promise may just never be fulfilled for whatever reason. If this happens, Promise.reject() is now called so that I have a clue about what's going on.