Make message about muting players appear in demos again...
If you play the demo back in the same game session after recording, you voice will still be muted. If you restart your game AND THEN play the demo, you will no longer be muted. For this purpose, I've ensured that the messages left in the console will be recorded still for DemoPlaybackHelper to pick up on. If you've restarted your game since recording the demo, it won't be necessary to have this tool running. This is only necessary if you watch the demo immediately after recording it.