Added test for Config
This is the start of the work for #23
Of note is the fact that the test directory is ignored in tsconfig.json. Without this, things were wonky and there would be a src and test directory inside the dist directory instead of dist containing the top level of files in src.