Fixed a significant blunder
When the ini library parses the ini file, all values come in as strings, even numbers. However, my tests were using a config structure with numeric values instead of strings. This makes it so that it no longer matters for config variables that are supposed to be numeric but could come in as a string at runtime or a number during tests.