Partial fix for #32
The console printing "Unknown command: If" was caused by a missing echo command in the message sent to the user in this case. Everything else about this branch of the logic may be functional. I never gave it enough time to see if it was properly working before I assumed there was an issue because of the command error printed to console. This should be merged into the main branch and then #32 should be monitored to see if recording is able to begin successfully after the round ends.