Fixed a crash!!!
Fixed a crash resulting from the changes in e2a0d3088529e3eaa8aeb50bd365abfff32b2aa5 and 28a7984268c42a86f6f4031e2de03c0da190ced6.This is because the subscribers list could have changed by the time it is referenced here. It's better to hold onto a reference of the desired listener to be able to access its properties until the end of the interaction.